pyproject.toml
[project]
name = "capitaloneassistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "faker>=37.12.0",
    "fastapi>=0.121.1",
    "google-generativeai>=0.8.5",
    "openai>=2.7.1",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "streamlit>=1.51.0",
    "uvicorn>=0.38.0",
]
